apollo_consensus_orchestrator: rename cende python_test module to python_compat
